moved to location_normalize() to location.py
